requirements.txt
python-telegram-bot==20.7
requests==2.31.0
flask==2.3.3